added additional users unit tests, reaching full coverage

+++ backend/src/controllers/users.controller.ts
@@ -1,17 +1,20 @@
 import { NextFunction, Request, Response } from "express";
+import { BadRequestError } from "../errors/index.js";
+import { isUser } from "../models/users.model.js";
 import { UsersService } from "../services/users.service.js";
 
-const usersService = new UsersService();
+const defaultUsersService = new UsersService();
 
 export const getUserById = async (
   req: Request,
   res: Response,
-  next: NextFunction
+  next: NextFunction,
+  usersService: UsersService = defaultUsersService
 ) => {
   try {
     const id = req.params["id"];
     if (!id) {
-      throw new Error(`Invalid id submitted`);
+      throw new BadRequestError(new Error(`Invalid id submitted`));
     }
     const user = await usersService.findById(id);
     res.json(user);
@@ -23,10 +26,16 @@ export const getUserById = async (
 export const updateUser = async (
   req: Request,
   res: Response,
-  next: NextFunction
+  next: NextFunction,
+  usersService: UsersService = defaultUsersService
 ) => {
   try {
     const userUpdates = req.body;
+    if (!isUser(userUpdates)) {
+      throw new BadRequestError(
+        new Error("Request body does not match user shape")
+      );
+    }
     const updatedUser = await usersService.update(userUpdates);
     res.json(updatedUser);
   } catch (err) {
@@ -37,10 +46,16 @@ export const updateUser = async (
 export const deleteUser = async (
   req: Request,
   res: Response,
-  next: NextFunction
+  next: NextFunction,
+  usersService: UsersService = defaultUsersService
 ) => {
   try {
     const userToDelete = req.body;
+    if (!isUser(userToDelete)) {
+      throw new BadRequestError(
+        new Error("Request body does not match user shape")
+      );
+    }
     const deletedUser = await usersService.delete(userToDelete);
     res.json(deletedUser);
   } catch (err) {
